Rename subprocess.environment to subprocesses.inspect_environment
